// src/composables/skill-learning/types.ts
export interface MelodyNote {
  note: string;
  duration: number;
  time: number;
}